@jbduncan 2024-03-06
Actually, the only code change here is in a test, so there's no need to run any benchmarks.