@jbduncan 2024-03-13
@kortschak I've run all the benchmarks for `dsp/fourier` on `master` vs this PR, and surprisingly there are some minor differences, but I cannot explain them as there were no changes to any benchmarked code, I ran as few programs as possible, and I ran the benchmarks with `-count=10`...

Anyway, I'm ready for a review when you are. :)

```
goos: darwin
goarch: arm64
pkg: gonum.org/v1/gonum/dsp/fourier
                               │ benchmarks_dsp_fourier_old.txt │   benchmarks_dsp_fourier_new.txt   │
                               │             sec/op             │   sec/op     vs base               │
RealFFTCoefficients/16-8                            146.3n ± 0%   146.4n ± 0%       ~ (p=0.132 n=10)
RealFFTCoefficients/128-8                           1.487µ ± 2%   1.487µ ± 1%       ~ (p=0.317 n=10)
RealFFTCoefficients/1024-8                          12.12µ ± 0%   12.09µ ± 0%  -0.22% (p=0.050 n=10)
RealFFTCoefficients/8192-8                          121.0µ ± 1%   121.0µ ± 0%       ~ (p=0.315 n=10)
RealFFTCoefficients/65536-8                         1.094m ± 1%   1.097m ± 0%       ~ (p=0.796 n=10)
RealFFTCoefficients/524288-8                        10.78m ± 1%   10.70m ± 0%  -0.79% (p=0.001 n=10)
RealFFTCoefficients/4194304-8                       102.8m ± 1%   101.8m ± 2%       ~ (p=0.218 n=10)
RealFFTCoefficients/100-8                           1.049µ ± 0%   1.047µ ± 0%  -0.24% (p=0.001 n=10)
RealFFTCoefficients/4000-8                          50.01µ ± 0%   49.65µ ± 0%  -0.72% (p=0.000 n=10)
RealFFTCoefficients/1000000-8                       19.02m ± 0%   18.91m ± 1%  -0.57% (p=0.029 n=10)
RealFFTSequence/16-8                                144.2n ± 0%   144.0n ± 0%       ~ (p=0.058 n=10)
RealFFTSequence/128-8                               1.625µ ± 0%   1.623µ ± 0%  -0.12% (p=0.001 n=10)
RealFFTSequence/1024-8                              13.70µ ± 0%   13.67µ ± 0%  -0.17% (p=0.016 n=10)
RealFFTSequence/8192-8                              139.6µ ± 0%   138.6µ ± 1%  -0.77% (p=0.002 n=10)
RealFFTSequence/65536-8                             1.269m ± 0%   1.262m ± 1%  -0.54% (p=0.000 n=10)
RealFFTSequence/524288-8                            12.44m ± 0%   12.43m ± 0%       ~ (p=0.143 n=10)
RealFFTSequence/4194304-8                           116.1m ± 0%   115.9m ± 0%  -0.13% (p=0.002 n=10)
RealFFTSequence/100-8                               1.157µ ± 0%   1.149µ ± 0%  -0.69% (p=0.000 n=10)
RealFFTSequence/4000-8                              58.21µ ± 0%   57.93µ ± 0%  -0.49% (p=0.000 n=10)
RealFFTSequence/1000000-8                           22.68m ± 0%   22.69m ± 0%       ~ (p=1.000 n=10)
CmplxFFTCoefficients/16-8                           208.6n ± 0%   207.0n ± 0%  -0.74% (p=0.001 n=10)
CmplxFFTCoefficients/128-8                          2.435µ ± 0%   2.417µ ± 0%  -0.76% (p=0.000 n=10)
CmplxFFTCoefficients/1024-8                         23.44µ ± 0%   23.29µ ± 0%  -0.66% (p=0.002 n=10)
CmplxFFTCoefficients/8192-8                         272.9µ ± 0%   271.7µ ± 0%  -0.41% (p=0.003 n=10)
CmplxFFTCoefficients/65536-8                        2.597m ± 0%   2.600m ± 2%       ~ (p=0.631 n=10)
CmplxFFTCoefficients/524288-8                       26.20m ± 0%   26.60m ± 1%  +1.53% (p=0.000 n=10)
CmplxFFTCoefficients/4194304-8                      250.4m ± 0%   251.6m ± 0%  +0.47% (p=0.000 n=10)
CmplxFFTCoefficients/100-8                          1.829µ ± 0%   1.830µ ± 0%       ~ (p=0.580 n=10)
CmplxFFTCoefficients/4000-8                         108.4µ ± 0%   108.3µ ± 1%       ~ (p=0.305 n=10)
CmplxFFTCoefficients/1000000-8                      42.96m ± 0%   42.92m ± 1%       ~ (p=0.143 n=10)
CmplxFFTSequence/16-8                               208.8n ± 0%   208.3n ± 0%  -0.24% (p=0.003 n=10)
CmplxFFTSequence/128-8                              2.437µ ± 0%   2.433µ ± 0%  -0.18% (p=0.001 n=10)
CmplxFFTSequence/1024-8                             23.47µ ± 0%   23.44µ ± 0%       ~ (p=0.148 n=10)
CmplxFFTSequence/8192-8                             272.8µ ± 0%   273.0µ ± 0%       ~ (p=0.353 n=10)
CmplxFFTSequence/65536-8                            2.596m ± 0%   2.627m ± 2%  +1.18% (p=0.000 n=10)
CmplxFFTSequence/524288-8                           26.20m ± 0%   26.39m ± 1%  +0.73% (p=0.000 n=10)
CmplxFFTSequence/4194304-8                          250.5m ± 0%   252.5m ± 0%  +0.79% (p=0.000 n=10)
CmplxFFTSequence/100-8                              1.830µ ± 0%   1.837µ ± 0%  +0.38% (p=0.000 n=10)
CmplxFFTSequence/4000-8                             108.4µ ± 0%   108.5µ ± 0%       ~ (p=0.971 n=10)
CmplxFFTSequence/1000000-8                          42.96m ± 0%   42.95m ± 0%  -0.03% (p=0.035 n=10)
Coefficients/Radix2/16-8                            73.36n ± 0%   73.27n ± 0%       ~ (p=0.210 n=10)
Coefficients/Radix4/16-8                            66.42n ± 0%   66.31n ± 0%       ~ (p=0.255 n=10)
Coefficients/Radix2/128-8                           828.2n ± 0%   829.6n ± 0%       ~ (p=0.183 n=10)
Coefficients/Radix2/1024-8                          10.21µ ± 0%   10.19µ ± 0%       ~ (p=0.118 n=10)
Coefficients/Radix4/1024-8                          6.901µ ± 0%   6.901µ ± 0%       ~ (p=0.565 n=10)
Coefficients/Radix2/8192-8                          115.2µ ± 0%   115.4µ ± 1%       ~ (p=0.481 n=10)
Coefficients/Radix2/65536-8                         1.429m ± 0%   1.429m ± 1%       ~ (p=0.529 n=10)
Coefficients/Radix4/65536-8                         859.6µ ± 1%   862.7µ ± 1%       ~ (p=0.247 n=10)
Coefficients/Radix2/524288-8                        13.89m ± 1%   13.86m ± 1%       ~ (p=0.123 n=10)
Coefficients/Radix2/4194304-8                       132.9m ± 1%   132.4m ± 1%       ~ (p=0.481 n=10)
Coefficients/Radix4/4194304-8                       82.79m ± 1%   83.96m ± 0%  +1.41% (p=0.000 n=10)
Sequence/Radix2/16-8                                85.23n ± 0%   85.61n ± 0%  +0.45% (p=0.003 n=10)
Sequence/Radix4/16-8                                76.54n ± 0%   76.53n ± 1%       ~ (p=0.670 n=10)
Sequence/Radix2/128-8                               889.4n ± 0%   888.9n ± 0%       ~ (p=0.224 n=10)
Sequence/Radix2/1024-8                              10.62µ ± 1%   10.61µ ± 0%       ~ (p=0.197 n=10)
Sequence/Radix4/1024-8                              7.307µ ± 0%   7.320µ ± 0%       ~ (p=0.424 n=10)
Sequence/Radix2/8192-8                              118.3µ ± 0%   118.1µ ± 0%       ~ (p=0.542 n=10)
Sequence/Radix2/65536-8                             1.466m ± 0%   1.459m ± 0%       ~ (p=0.052 n=10)
Sequence/Radix4/65536-8                             893.5µ ± 1%   891.4µ ± 1%       ~ (p=0.853 n=10)
Sequence/Radix2/524288-8                            14.23m ± 1%   14.11m ± 1%  -0.81% (p=0.002 n=10)
Sequence/Radix2/4194304-8                           135.7m ± 1%   135.3m ± 1%       ~ (p=0.280 n=10)
Sequence/Radix4/4194304-8                           85.00m ± 0%   86.36m ± 1%  +1.59% (p=0.000 n=10)
geomean                                             139.1µ        139.0µ       -0.05%
```